backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
mindsdb-sdk
sqlalchemy==2.0.27 